refact: router structure
