+++ ui/src/App.tsx
@@ -1,8 +1,8 @@
 import { UserProvider } from "./contexts/userContext";
+import { router } from "./router";
 import "./styles/global.css"
 
 import { RouterProvider } from "react-router-dom";
-import router from "./router";
 import { ToastContainer } from "react-toastify";
