+++ ui/src/components/Header/Menu/index.tsx
@@ -1,6 +1,8 @@
 import { MouseEventHandler } from "react";
 import styles from "./styles.module.css"
 import Link from "../../Link";
+import authenticatedRoutes from "../../../router/protected/authenticated.routes";
+import teacherRoutes from "../../../router/protected/teacher.routes";
 
 interface IMenuProps {
     open: boolean;
@@ -28,9 +30,13 @@ export default ({ open, handleClose }:IMenuProps) => {
                 >X</button>
 
                 <div className={styles.link_list}>
-                    {/* {Object.values(routesByAccessLevel).map((route, i) => (
+                    {Object.values(authenticatedRoutes).map((route, i) => (
                         <Link key={i} to={route.path}>{ route.title }</Link>
-                    ))} */}
+                    ))}
+
+                    {Object.values(teacherRoutes).map((route, i) => (
+                        <Link key={i} to={route.path}>{ route.title }</Link>
+                    ))}
                 </div>
             </div>
         </div>
