+++ ui/src/router/protected/authenticated.routes.tsx
